requirements.txt
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
decorator==5.2.1
h11==0.16.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.3.5
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
proglog==0.1.12
pycparser==2.23
pydub==0.25.1
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
selenium==4.38.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
